@@ -1,23 +1,21 @@
-package exercises;
+package answers;
 
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static io.restassured.RestAssured.given;
 
-public class WireMockExercises1 {
+@WireMockTest(httpPort = 9876)
+public class WireMockAnswers1Test {
 
 	private RequestSpecification requestSpec;
 
-	@Rule
-	public WireMockRule wireMockRule = new WireMockRule(9876);
-
-	@Before
+	@BeforeEach
 	public void createRequestSpec() {
 
 		requestSpec = new RequestSpecBuilder().
@@ -33,7 +31,10 @@ public class WireMockExercises1 {
 		 * to /pl/80-862 with an HTTP status code 200
 		 ************************************************/
 
-
+		stubFor(post(urlEqualTo("/pl/80-862"))
+				.willReturn(aResponse()
+						.withStatus(200)
+				));
 	}
 	
 	public void setupStubExercise102() {
@@ -43,7 +44,11 @@ public class WireMockExercises1 {
 		 * to /pl/80-863 with a response that contains
 		 * a Content-Type header with value text/plain
 		 ************************************************/
-
+		
+		stubFor(post(urlEqualTo("/pl/80-863"))
+				.willReturn(aResponse()
+						.withHeader("Content-Type", "text/plain")
+				));
 	}
 	
 	public void setupStubExercise103() {
@@ -53,6 +58,10 @@ public class WireMockExercises1 {
 		 * to /pl/80-864 with a response body 'Posted!'
 		 ************************************************/
 
+		stubFor(post(urlEqualTo("/pl/80-864"))
+				.willReturn(aResponse()
+						.withBody("Posted!")
+				));
 	}
 
 	@Test
