@@ -1,23 +1,23 @@
-package exercises;
+package answers;
 
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import com.github.tomakehurst.wiremock.http.Fault;
+import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
 import org.apache.http.client.ClientProtocolException;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static io.restassured.RestAssured.given;
 
-public class WireMockExercises2 {
+@WireMockTest(httpPort = 9876)
+public class WireMockAnswers2Test {
 
     private RequestSpecification requestSpec;
 
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(9876);
-
-    @Before
+    @BeforeEach
     public void createRequestSpec() {
 
         requestSpec = new RequestSpecBuilder().
@@ -35,6 +35,11 @@ public class WireMockExercises2 {
          * equal to 'Service unavailable'
          ************************************************/
 
+        stubFor(get(urlEqualTo("/servicedown"))
+            .willReturn(aResponse()
+                .withStatus(503)
+                .withStatusMessage("Service unavailable")
+            ));
     }
 
     public void setupStubExercise202() {
@@ -46,6 +51,12 @@ public class WireMockExercises2 {
          * fixed delay of 3000 milliseconds.
          ************************************************/
 
+        stubFor(get(urlEqualTo("/slow"))
+            .withHeader("speed", equalTo("slow"))
+            .willReturn(aResponse()
+                .withStatus(200)
+                .withFixedDelay(3000)
+            ));
     }
 
     public void setupStubExercise203() {
@@ -55,6 +66,10 @@ public class WireMockExercises2 {
          * to /fault with a Fault of type RANDOM_DATA_THEN_CLOSE
          ************************************************/
 
+        stubFor(get(urlEqualTo("/fault"))
+                .willReturn(aResponse()
+                        .withFault(Fault.RANDOM_DATA_THEN_CLOSE)
+                ));
     }
 
     @Test
@@ -99,7 +114,7 @@ public class WireMockExercises2 {
             time(org.hamcrest.Matchers.greaterThan(3000L));
     }
 
-    @Test(expected = ClientProtocolException.class)
+    @Test
     public void testExercise203() {
 
         /***
@@ -108,9 +123,12 @@ public class WireMockExercises2 {
 
         setupStubExercise203();
 
-        given().
-            spec(requestSpec).
-        when().
-            get("/fault");
+        Assertions.assertThrows(ClientProtocolException.class, () -> {
+
+            given().
+                spec(requestSpec).
+            when().
+                get("/fault");
+        });
     }
 }
