@@ -1,29 +1,30 @@
 package examples.usingextensions;
 
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.github.tomakehurst.wiremock.extension.Parameters;
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import examples.extensions.BasicAuthRequestFilter;
+import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import examples.extensions.BodyLengthMatcher;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static io.restassured.RestAssured.given;
 
-public class CustomMatcher {
+public class CustomMatcherTest {
 
     private RequestSpecification requestSpec;
 
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(
-            new WireMockConfiguration().port(9876).extensions(new BodyLengthMatcher())
-    );
+    @RegisterExtension
+    static WireMockExtension wiremock = WireMockExtension.newInstance().
+            options(wireMockConfig().
+                    port(9876).
+                    extensions(new BodyLengthMatcher())
+            ).build();
 
-    @Before
+    @BeforeEach
     public void createRequestSpec() {
 
         requestSpec = new RequestSpecBuilder().
@@ -34,11 +35,11 @@ public class CustomMatcher {
 
     public void stubForCustomMatching() {
 
-        stubFor(get(urlEqualTo("/custom-matching")).
+        wiremock.stubFor(get(urlEqualTo("/custom-matching")).
                 willReturn(aResponse().withStatus(200))
         );
 
-        stubFor(get(urlEqualTo("/custom-matching")).
+        wiremock.stubFor(get(urlEqualTo("/custom-matching")).
                 andMatching("body-too-long", Parameters.one("maxLength", 20)).
                 willReturn(aResponse().withStatus(400))
         );
