@@ -1,31 +1,33 @@
 package examples.usingextensions;
 
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import examples.extensions.CreateDateHeaderDefinitionTransformer;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static io.restassured.RestAssured.given;
 
-public class ResponseDefinitionTransformer {
+public class ResponseDefinitionTransformerTest {
 
     private RequestSpecification requestSpec;
 
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(
-            new WireMockConfiguration().port(9876).extensions(new CreateDateHeaderDefinitionTransformer())
-    );
+    @RegisterExtension
+    static WireMockExtension wiremock = WireMockExtension.newInstance().
+            options(wireMockConfig().
+                    port(9876).
+                    extensions(new CreateDateHeaderDefinitionTransformer())
+            ).build();
 
-    @Before
+    @BeforeEach
     public void createRequestSpec() {
 
         requestSpec = new RequestSpecBuilder().
@@ -36,7 +38,7 @@ public class ResponseDefinitionTransformer {
 
     public void stubForResponseDefinitionTransformer() {
 
-        stubFor(get(urlEqualTo("/response-definition-transformer"))
+        wiremock.stubFor(get(urlEqualTo("/response-definition-transformer"))
                 .willReturn(aResponse()
                         .withTransformerParameter("dateFormat", "dd-MM-yyyy")
                 ));
