@@ -1,13 +1,13 @@
 package exercises;
 
+import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import exercises.extensions.MultipleHttpVerbsMatcher;
 import com.github.tomakehurst.wiremock.extension.Parameters;
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,18 +16,18 @@ import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static io.restassured.RestAssured.given;
 
-public class WireMockExercises5dot2 {
+public class WireMockExercises5dot2Test {
 
     private RequestSpecification requestSpec;
 
-    @Rule
-    public WireMockRule wireMockRule =
-        new WireMockRule(wireMockConfig().
-            port(9876).
-            extensions(new MultipleHttpVerbsMatcher())
-        );
+    @RegisterExtension
+    static WireMockExtension wiremock = WireMockExtension.newInstance().
+            options(wireMockConfig().
+                    port(9876).
+                    extensions(new MultipleHttpVerbsMatcher())
+            ).build();
 
-    @Before
+    @BeforeEach
     public void createRequestSpec() {
 
         requestSpec = new RequestSpecBuilder().
@@ -42,7 +42,7 @@ public class WireMockExercises5dot2 {
         verbs.add("GET");
         verbs.add("POST");
 
-        stubFor(requestMatching("multiple-verbs-matcher",
+        wiremock.stubFor(requestMatching("multiple-verbs-matcher",
                         Parameters.one("acceptedVerbs", verbs))
             .willReturn(aResponse()
                 .withStatus(200)
@@ -90,5 +90,24 @@ public class WireMockExercises5dot2 {
             statusCode(200).
         and().
             body(org.hamcrest.Matchers.equalTo("Your HTTP verb has been accepted!"));
+    }
+
+    @Test
+    public void anIncomingHttpPATCHShouldNotBeMatched() {
+
+        /***
+         * Use this test to test your implementation of the custom matcher
+         * This request should not be accepted as it uses an HTTP PATCH
+         */
+
+        setupStubExercise5dot2();
+
+        given().
+            spec(requestSpec).
+        when().
+            patch("/match-multiple-verbs").
+        then().
+            assertThat().
+            statusCode(404);
     }
 }
