@@ -1,22 +1,20 @@
-package answers;
+package exercises;
 
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
-import org.junit.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static io.restassured.RestAssured.given;
 
-public class WireMockAnswers1 {
+@WireMockTest(httpPort = 9876)
+public class WireMockExercises1Test {
 
 	private RequestSpecification requestSpec;
 
-	@Rule
-	public WireMockRule wireMockRule = new WireMockRule(9876);
-
-	@Before
+	@BeforeEach
 	public void createRequestSpec() {
 
 		requestSpec = new RequestSpecBuilder().
@@ -32,10 +30,7 @@ public class WireMockAnswers1 {
 		 * to /pl/80-862 with an HTTP status code 200
 		 ************************************************/
 
-		stubFor(post(urlEqualTo("/pl/80-862"))
-				.willReturn(aResponse()
-						.withStatus(200)
-				));
+
 	}
 	
 	public void setupStubExercise102() {
@@ -45,11 +40,7 @@ public class WireMockAnswers1 {
 		 * to /pl/80-863 with a response that contains
 		 * a Content-Type header with value text/plain
 		 ************************************************/
-		
-		stubFor(post(urlEqualTo("/pl/80-863"))
-				.willReturn(aResponse()
-						.withHeader("Content-Type", "text/plain")
-				));
+
 	}
 	
 	public void setupStubExercise103() {
@@ -59,10 +50,6 @@ public class WireMockAnswers1 {
 		 * to /pl/80-864 with a response body 'Posted!'
 		 ************************************************/
 
-		stubFor(post(urlEqualTo("/pl/80-864"))
-				.willReturn(aResponse()
-						.withBody("Posted!")
-				));
 	}
 
 	@Test
