@@ -1,22 +1,19 @@
-package answers;
+package exercises;
 
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import com.github.tomakehurst.wiremock.stubbing.Scenario;
+import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
-import org.junit.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static io.restassured.RestAssured.given;
 
-public class WireMockAnswers3 {
+@WireMockTest(httpPort = 9876)
+public class WireMockExercises3Test {
 
     private RequestSpecification requestSpec;
 
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(9876);
-
-    @Before
+    @BeforeEach
     public void createRequestSpec() {
 
         requestSpec = new RequestSpecBuilder().
@@ -37,32 +34,6 @@ public class WireMockAnswers3 {
          *      returns HTTP 200 and body "DATA FOR /nl/3825"
          ************************************************/
 
-        stubFor(get(urlEqualTo("/nl/3825"))
-            .inScenario("Stateful mock exercise")
-            .whenScenarioStateIs(Scenario.STARTED)
-            .willReturn(aResponse()
-                .withStatus(404)
-            )
-        );
-
-        stubFor(post(urlEqualTo("/nl/3825"))
-            .inScenario("Stateful mock exercise")
-            .whenScenarioStateIs(Scenario.STARTED)
-            .withRequestBody(equalTo("DATA FOR /nl/3825"))
-            .willReturn(aResponse()
-                .withStatus(201)
-            )
-            .willSetStateTo("DATA_CREATED")
-        );
-
-        stubFor(get(urlEqualTo("/nl/3825"))
-            .inScenario("Stateful mock exercise")
-            .whenScenarioStateIs("DATA_CREATED")
-            .willReturn(aResponse()
-                .withStatus(200)
-                .withBody("DATA FOR /nl/3825")
-            )
-        );
     }
 
     @Test
